requirements.txt
aiosqlite==0.17.0
alembic==1.8.1
ansi2html==1.7.0
anyio==3.6.1
api-client==1.3.1
apiclient==1.0.4
asgiref==3.5.1
assertpy==1.1
asttokens==2.0.5
async-generator==1.10
asyncio==3.4.3
asyncpg==0.25.0
atomicwrites==1.4.0
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
chai==1.1.2
charset-normalizer==2.0.12
click==8.1.3
cloudinary==1.30.0
colorama==0.4.4
constantly==15.1.0
cryptocode==0.1
cryptography==37.0.4
databases==0.6.0
decopatch==1.4.10
decorator==5.1.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.1
dnspython==2.2.1
docutils==0.18.1
ecdsa==0.17.0
email-validator==1.2.1
et-xmlfile==1.1.0
executing==0.8.2
Faker==13.3.3
fastapi==0.77.1
FastAPI-SQLModel==0.0.1
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.13.0
hasher==2.0.1
heartrate==0.2.2
htmlmin==0.1.12
httptools==0.4.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.8.0
incremental==21.3.0
iniconfig==1.1.1
is-valid-date==0.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
json3==1.0
jsoncomparedeep==1.20
jwt==1.3.1
logging2==0.1.2
loguru==0.5.3
makefun==1.14.0
Mako==1.2.1
MarkupSafe==2.1.1
numpy==1.23.0
openpyxl==3.0.10
orjson==3.6.8
outcome==1.1.0
packaging==21.3
panda==0.3.1
pandas==1.4.3
passlib==1.7.4
Password-validate==1.0.0
pcrypt==1.0.5
phonenumbers==8.12.41
Pillow==9.0.1
pipenv==2022.5.2
platformdirs==2.5.2
playwright==1.23.0
pluggy==1.0.0
psycopg==3.0.12
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycountry==22.1.10
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.9.0
pyee==8.1.0
Pygments==2.11.2
PyJWT==2.4.0
pyparsing==3.0.7
pytest==7.1.1
pytest-asyncio==0.19.0
pytest-base-url==1.4.2
pytest-cases==3.6.13
pytest-html==3.1.1
pytest-html-reporter==0.2.9
pytest-metadata==2.0.2
pytest-playwright==0.3.0
pytest-reporter==0.5.2
pytest-reporter-html1==0.8.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==6.1.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlalchemy2-stubs==0.0.2a22
SQLAlchemy==1.4.31
sqlmodel==0.0.8
sqlparse==0.4.2
starlette==0.19.1
tenacity==8.0.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.1
trio==0.20.0
Twisted==22.2.0
twisted-iocpsupport==1.0.2
typing==3.7.4.3
typing_extensions==4.2.0
tzdata==2022.1
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
validator==0.7.1
virtualenv==20.14.1
virtualenv-clone==0.5.7
watchgod==0.8.2
websockets==10.1
Werkzeug==2.0.3
whitenoise==5.3.0
win32-setctime==1.0.4
zipp==3.8.0
zope.interface==5.4.0
